Use string SharedPreferences key for update ringtone.
Instead of using the preference, deal directly with the key.

I changed the logic so that it will return null/silent as the
ringtone uri if the key is null. Because, technically, one could
call the function with a "type" expecting a valid return type,
but not provide a key.

Bug: 18232725
Change-Id: Ifb3d517ef8b0a5ddba388f2e0d471874f7b4b235