Backport Context.getDrawable()
Context.getDrawable() was added in API level 21. In order to backport it,
we could use ContextCompat.getDrawable(). However, the code in PhoneCommon
is built as part of master in packages/services/Telephony as well and it
doesn't include the support library. Simply importing ContextCompat would
break master, so we handle compatibility in the class itself.

Bug: 25629359
Change-Id: I98a413a1b7e0a9a94a722059e18318e22939040d