Add ability to disable auto-resize of text for resizing text view.
The emergency dialer uses this component.  We will sometimes show a
static status hint in the dialpad edit text; in that case we want to just
set the size of the text to a fixed size and disable the auto-resize.

Test: Manual
Bug: 68030294
Merged-In: I6f8ec8851874963afac66fcd60812d6ab2fec4ba
Change-Id: I6f8ec8851874963afac66fcd60812d6ab2fec4ba